+++ brue/js_reserves.py
@@ -1,8 +1,16 @@
 # -*- coding: utf-8 -*-
 
-class _This(object):
-    """
-    javascript this reserved words
-    """
+class FileReader:
+    @property
+    def result(self) -> str:
+        return ""
 
-this = _This()
+    def onchange(self):
+        pass
+
+class JSON:
+    def parse(self, source:str) -> dict:
+        return {}
+
+    def stringify(self, source:dict) -> str:
+        return ""